updated request page - still in progress
