+++ src/test/java/ProcessBuilderTest.java
@@ -1,7 +1,9 @@
+import Utils.Log;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.annotations.Test;
+import pageObjects.BasePage;
 import pageObjects.LoginPage;
 import pageObjects.ProcessBuilderPage;
 import pageObjects.SalesforceSetupPage;
@@ -11,22 +13,28 @@ public class ProcessBuilderTest {
     String password= System.getProperty("password", "local").trim();
     String sandBoxURL= System.getProperty("sandboxURL", "local").trim();
 
+    public static String Requestor_Admin = "Requestor_Admin";
+    public static String Responder_Admin = "Responder_Admin";
+    public static String Laboratory_Admin = "Laboratory_Admin";
+    public static String Requestor_SPU = "Requestor_SPU";
+    public static String Responder_SPU = "Responder_SPU";
+    public static String Laboratory_SPU = "Laboratory_SPU";
+
+    public static String baseUrl = "https://login.salesforce.com";
+
     @Test
-    public void navigateToProcessBuilder()  {
+    public void navigateToProcessBuilder() throws Exception {
         WebDriver driver = doLogin();
 
     }
 
-    private WebDriver doLogin(){
+    private WebDriver doLogin() throws Exception {
         WebDriver driver = new ChromeDriver();
 
         LoginPage loginPage = new LoginPage(driver);
-        loginPage.navigateToPage(sandBoxURL);
-
-        loginPage.getUsernameInput().sendKeys(username);
-        loginPage.getPasswordInput().sendKeys(password);
-        loginPage.getLoginButton().click();
-
+        loginPage.navigateToPage(baseUrl);
+//        loginPage.loginAs(Requestor_Admin);
+        loginPage.LoginUser("username", "password");
         return driver;
     }
 
@@ -35,7 +43,7 @@ public class ProcessBuilderTest {
         return random;
     }
 
-    private void oldTest() throws InterruptedException{
+    private void oldTest() throws Exception {
         WebDriver driver = doLogin();
 
         //Navigation to process builder
